Constantin: use compile-time env if possible
Constantin reads the NOOP_HOME env dynamically during simulation.
This requires the env to be always set during simulation time.
However, this requirement seems useless as the user seldom changes
the env path.

This commit adds the compile-time NOOP_HOME variable as the fallback
option if the env variable is not set during simulation time.