fix(AsyncBridge): set `tx/rxlinkactivereq` HIGH only when bridge is ready
When the asynchronous bridge receives rx*lcrdv not long after reset, the
bridge is not yet ready because the sink part must wait for `ridx` of
the source part to be valid in the safe mode. Therefore rx*lcrdv is lost
while the sink part is still not ready. The same is true for tx*lcrdv in
the opposite direction. To fix this bug, this commit sets
`tx/rxlinkactivereq` HIGH only when reset process is finish in both
source and sink part of the bridge.