Implement read-write conflict behavior in SRAM template
This setting determines the conflict behavior of the SRAMTemplate module as seen
by connected logic, as well as the expected conflict behavior of the underlying
RAM macro used within the module. For example, when the underlying RAM macro
does not support read-write conflicts, the SRAMTemplate module can insert some
additional logic around the macro which provides limited support.